backend/requirements.txt
fastapi
uvicorn
httpx
yfinance
feedparser
pytz
supabase
python-dotenv
pandas
numpy
pydantic
requests
jugaad-data
nsetools
openpyxl
lxml
html5lib
beautifulsoup4
pypdf
pytest
